Add support for GlobalSign CA
The npm website is now using a GlobalSign certificate.  Though it
pains me greatly to reduce security by using a CA for the npm client
that is not used for only that purpose, it would be very convenient
to be able to have people load the registry in their web browsers
without telling me that it's a 'self-signed' certificate (which is
patently untrue, of course.)

I have decided to stop answering the call of duty, and instead,
allow people to be wrong on the internet.
http://xkcd.com/386/

However, to make this work, we have to continue supporting the npmCA
cert for a while, and only switch over once enough people have
upgraded to a version of npm that supports the GlobalSign cert.